Refactor dev tunnel tests into separate file.
